Lots of UI adjustments
