+++ src/api/java/com/minecolonies/api/util/SoundUtils.java
@@ -203,6 +203,10 @@ public final class SoundUtils
               (float) 1.0,
               player.level().random.nextLong()));
         }
+        else
+        {
+            player.playNotifySound(SoundEvents.NOTE_BLOCK_BELL.get(), SoundSource.NEUTRAL, 1.0f, 1.0f);
+        }
     }
 
     /**
@@ -222,6 +226,10 @@ public final class SoundUtils
               (float) VOLUME * 2,
               (float) 0.3,
               player.level().random.nextLong()));
+        }
+        else
+        {
+            player.playNotifySound(SoundEvents.NOTE_BLOCK_DIDGERIDOO.get(), SoundSource.NEUTRAL, 1.0f, 0.3f);
         }
     }
