+++ src/main/java/com/minecolonies/coremod/client/gui/AbstractModuleWindow.java
@@ -6,8 +6,10 @@ import com.minecolonies.api.colony.buildings.modules.IBuildingModuleView;
 import com.minecolonies.api.colony.buildings.modules.IModuleWindow;
 import com.minecolonies.api.colony.buildings.views.IBuildingView;
 import com.minecolonies.coremod.colony.buildings.views.AbstractBuildingView;
+import net.minecraft.client.resources.sounds.SimpleSoundInstance;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.sounds.SoundEvents;
 
 import java.util.Locale;
 import java.util.Random;
@@ -69,7 +71,10 @@ public abstract class AbstractModuleWindow extends AbstractWindowSkeleton implem
             image.setImage(new ResourceLocation("minecolonies:textures/gui/modules/tab_side" + (random.nextInt(3) + 1) + ".png"), false);
             image.setPosition(-20, 10 + offset);
             image.setSize(32, 26);
-            image.setHandler(button -> view.getWindow().open());
+            image.setHandler(button -> {
+                mc.getSoundManager().play(SimpleSoundInstance.forUI(SoundEvents.BOOK_PAGE_TURN, 1.0F));
+                view.getWindow().open();
+            });
 
             final String icon = view.getIcon();
             final ButtonImage iconImage = new ButtonImage();
@@ -77,7 +82,10 @@ public abstract class AbstractModuleWindow extends AbstractWindowSkeleton implem
             iconImage.setSize(20, 20);
             iconImage.setID(icon);
             iconImage.setPosition(-15, 13 + offset);
-            iconImage.setHandler(button -> view.getWindow().open());
+            iconImage.setHandler(button -> {
+                mc.getSoundManager().play(SimpleSoundInstance.forUI(SoundEvents.BOOK_PAGE_TURN, 1.0F));
+                view.getWindow().open();
+            });
 
             offset += image.getHeight() + 2;
