+++ src/main/java/com/minecolonies/coremod/network/messages/PermissionsMessage.java
@@ -330,7 +330,6 @@ public class PermissionsMessage
                 colony.getPermissions().addRank(rankName);
             }
         }
-
     }
 
     /**
