--- src/main/resources/assets/minecolonies/gui/townhall/layoutsettings.xml
@@ -1,75 +1,0 @@
-<window size="524 243" pause="false" lightbox="false">
-
-    <layout source="minecolonies:gui/townhall/windowtownhall.xml"/>
-    <label id="assignCitizens" label="$(com.minecolonies.coremod.gui.townhall.jobhiring)" pos="100 62" size="148 11" textalign="MIDDLE" color="black"/>
-    <button id="toggleJob" size="57 20" pos="145 75" label="$(com.minecolonies.coremod.gui.hiring.off)"/>
-
-    <label id="assignHouses" label="$(com.minecolonies.coremod.gui.townhall.housing)" pos="100 102" size="148 11" textalign="MIDDLE" color="black"/>
-    <button id="toggleHousing" size="57 20" pos="145 115" label="$(com.minecolonies.coremod.gui.hiring.off)"/>
-
-    <label id="movingIn" label="$(com.minecolonies.coremod.gui.townhall.movein)" pos="100 182" size="148 11" textalign="MIDDLE" color="black"/>
-    <button id="toggleMoveIn" size="57 20" pos="145 195" label="$(com.minecolonies.coremod.gui.townhall.off)"/>
-
-    <label id="printProgress" label="$(com.minecolonies.coremod.gui.townhall.printprogress)" pos="100 142" size="148 11" textalign="MIDDLE" color="black"/>
-    <button id="togglePrintProgress" size="57 20" pos="145 155" label="$(com.minecolonies.coremod.gui.townhall.on)"/>
-
-    <label id="pickColor" label="$(com.minecolonies.coremod.gui.townhall.pickteamcolor)" pos="276 28" size="148 11" textalign="MIDDLE" color="black"/>
-    <buttonimage id="previousColor" size="14 15" pos="292 42" label="&lt;" source="minecolonies:textures/gui/builderhut/builder_button_mini.png" textcolor="black"/>
-    <dropdown id="colorPicker"
-              size="86 17"
-              pos="306 42"
-              maxContentHeight="75"
-              dropDownSize="86 75"
-              source="minecolonies:textures/gui/builderhut/builder_button_medium.png"
-              textcolor="black"
-              scrollbarOffset="9 0"
-              dropfixx="75">
-        <view size="100% 15">
-            <buttonimage id="button" size="86 17" pos="0 0" source="minecolonies:textures/gui/builderhut/builder_button_medium.png" textcolor="black"/>
-        </view>
-    </dropdown>
-    <buttonimage id="nextColor" size="14 15" pos="392 42" label="&gt;" source="minecolonies:textures/gui/builderhut/builder_button_mini.png" textcolor="black"/>
-    <button id="bannerPicker" label="$(com.minecolonies.coremod.gui.townhall.edit_flag)" pos="292 70" size="114 20"/>
-
-    <label label="$(com.minecolonies.coremod.gui.townhall.picktexturestyle)" pos="276 100" size="148 11" textalign="MIDDLE" color="black"/>
-    <dropdown id="textureStylePicker"
-              size="86 17"
-              pos="306 114"
-              maxContentHeight="75"
-              dropDownSize="86 75"
-              source="minecolonies:textures/gui/builderhut/builder_button_medium.png"
-              disabled="minecolonies:textures/gui/builderhut/builder_button_medium_dark.png"
-              textcolor="black"
-              scrollbarOffset="9 0"
-              dropfixx="75"
-              enabled="false">
-        <view size="100% 15">
-            <buttonimage id="button" size="86 17" pos="0 0" source="minecolonies:textures/gui/builderhut/builder_button_medium.png" textcolor="black"/>
-        </view>
-    </dropdown>
-
-    <label label="$(com.minecolonies.coremod.gui.townhall.pickcolonystyle)" pos="276 140" size="148 11" textalign="MIDDLE" color="black"/>
-    <buttonimage id="colonyStylePicker"
-              size="86 17"
-              pos="306 154"
-              source="minecolonies:textures/gui/builderhut/builder_button_medium.png"
-              disabled="minecolonies:textures/gui/builderhut/builder_button_medium_dark.png"
-              textcolor="black"/>
-
-    <label label="$(com.minecolonies.coremod.gui.townhall.picknamestyle)" pos="276 180" size="148 11" textalign="MIDDLE" color="black"/>
-    <dropdown id="nameStylePicker"
-              size="86 17"
-              pos="306 194"
-              maxContentHeight="30"
-              dropDownSize="86 30"
-              source="minecolonies:textures/gui/builderhut/builder_button_medium.png"
-              disabled="minecolonies:textures/gui/builderhut/builder_button_medium_dark.png"
-              textcolor="black"
-              scrollbarOffset="9 0"
-              dropfixx="30"
-              enabled="false">
-        <view size="100% 15">
-            <buttonimage id="button" size="86 17" pos="0 0" source="minecolonies:textures/gui/builderhut/builder_button_medium.png" textcolor="black"/>
-        </view>
-    </dropdown>
-</window>
