+++ src/main/resources/assets/minecolonies/gui/townhall/layoutstats.xml
@@ -11,7 +11,7 @@
             <buttonimage id="button" size="129 17" pos="0 0" source="minecolonies:textures/gui/builderhut/builder_button_medium_large.png" textcolor="black"
                      label="$(com.minecolonies.coremod.gui.interval.yesterday)"/>
         </dropdown>
-        <list id="stats" size="160 145" pos="20 80">
+        <list id="stats" size="160 135" pos="20 85">
            <label id="desc" size="160 13" pos="0 0" color="black"> </label>
         </list>
     </view>
